separate c++ class for better readability, add GetPingServiceEndpoints
